# docs/pg-del-stats-swagger.yaml
swagger: '2.0'
info:
  description: Programmatic Gauranteed Delivery Stats API
  version: 1.0.0
  title: PG Del Stats API

  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html

# tags are used to represent authentication roles
...
paths:
...
  /v1/report/token-spend:
    get:
      tags:
      - gp
      summary: retrieve Token Spend reports
      operationId: retrieveTokenSpendReportV1
      description: |
        Set of all token spend reports aggregated since last request
        or since a requested time.
      produces:
      - application/json
      parameters:
      - in: query
        name: since
        description: |
          Return token spend reports since this time
          (up to 15 mins from current time)
        type: string
        format: date-time
        required: false
      - in: query
        name: vendor
        description: Vendor code, used to filter reports returned
        required: false
        type: string
      - in: query
        name: region
        description: Region, used to filter reports returned
        required: false
        type: string
      responses:
        200:
          description: |
            Set of all token spend reports aggregated since last
            request or since requested time if provided.
          schema:
            $ref: '#/definitions/TokenSpendResponse'
...
definitions:
...
  TokenSpendResponse:
    type: object
    required:
    - tokenSpendSummaryLines
    properties:
      tokenSpendSummaryLines:
        type: array
        items:
          $ref: '#/definitions/TokenSpendSummary'
  TokenSpendSummary:
    type: object
    required:
    - reportId
    - vendor
    - region
    - instanceId
    - bidderCode
    - lineItemId
    - extLineItemId
    - dataWindowStartTimestamp
    - dateWindowEndTimestamp
    - reportTimestamp
    - serviceInstanceId
    - summaryData
    properties:
      reportId:
        type: string
        format: uuid
        example: 5996ee56-60c5-430f-ae48-57f1304443f7
      vendor:
        type: string
        example: "bidder"
      region:
        type: string
        example: "us-west-2"
      instanceId:
        type: string
        example: "18374988a98649bc8460345f066e4bf2"
      bidderCode:
        type: string
        example: "bidderPG"
      lineItemId:
        type: string
        example: "bidderPG-2251"
      extLineItemId:
        type: string
        example: "2251"
      dataWindowStartTimestamp:
        type: string
        format: date-time
        example: "2020-08-28T20:41:00.000Z"
      dataWindowEndTimestamp:
        type: string
        format: date-time
        example: "2020-08-28T20:42:00.000Z"
      reportTimestamp:
        type: string
        format: date-time
        example: "2020-08-28T20:42:00.000Z"
      serviceInstanceId:
        type: string
        example: "instance-mbp"
      summaryData:
        type: object
        required:
        - targetMatched
        properties:
          targetMatched:
            type: integer
            example: 0

# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /Forced-Team/PG-Del-Stats/1.0.0
schemes:
 - https